Add new requirement (add rating to book)

+++ src/main/java/com/spacecodee/library_book_backend/model/vo/rating/book/RatingBookVo.java
@@ -16,5 +16,5 @@ import java.io.Serializable;
 public class RatingBookVo implements Serializable {
     private int clientId;
     private int bookId;
-    private short ratingBook;
+    private double ratingBook;
 }
